refactor: move unsupported builtins into a new file
